New InkPen tool and start of the tool options bar.
InkPen kinda blots the stroke as you go slower (translation: the width
of strokes between nodes gets thinner as their distance increases). It's
pretty satisfying to draw with, for some reason.
Tool options bar is underway with selectors for color (note to self:
need to simplify the pop-up dialog) and a width selector. I kinda want
this pane, in the end, to look like GIMP's, so that's the design I have
in mind.
Next I need to make the animation view, that'll be fun.